feat: 添加current
re #89

+++ src/components/d-left-scene/index.vue
@@ -5,19 +5,23 @@
 		span 场景区
 	header.fn-flex.flex-row
 		i-input(
-			:value="editor.sceneObj[editor.sceneIndex].name",
+			:value="editor.sceneObj[editor.currentSceneIndex].name",
 			@on-change="handleSceneName",
 			@on-focus="event.inputFocus = true",
 			@on-blur="event.inputFocus = false",
 			v-if="editScene")
 			i-icon(type="md-checkmark", slot="suffix", @click="editScene = false")
-		i-select(v-model="editor.sceneIndex", v-if="!editScene", filterable)
+		i-select(
+			:value="editor.currentSceneIndex",
+			v-if="!editScene",
+			filterable,
+			@on-change="changeSceneIndex")
 			i-option(:value="0") 主场景
 			i-option(:value="key", v-for="(item, key) in editor.sceneObj", :key="key") {{ item.name }}
 			i-option(:value="-1") 回收站
 	ul.d-scrollbar.d-left-scene-list
 		draggable(
-			v-model="editor.sortByZIndexWidgetsList",
+			:value="editor.sortByZIndexWidgetsList",
 			@change="sceneWidgetDragEnd")
 			transition-group
 				item-card(
@@ -29,15 +33,15 @@
 		.d-left-scene-bottom-btn.text-center(@click="handleSetScene('create')") 新增
 		.d-left-scene-bottom-btn.text-center(
 			@click="handleSetScene('edit')",
-			:class="{ disabled: editor.sceneIndex === 0 }") 编辑
+			:class="{ disabled: editor.currentSceneIndex === 0 }") 编辑
 		.d-left-scene-bottom-btn.text-center(
 			@click="handleSetScene('destroy')",
-			:class="{ disabled: editor.sceneIndex === 0 }") 删除
+			:class="{ disabled: editor.currentSceneIndex === 0 }") 删除
 	i-modal(v-model="copyModel", title="提示", :footer-hide="true")
 		.fn-flex.flex-row
-			span.fn-hide.copy-id {{ editor.sceneIndex }}
+			span.fn-hide.copy-id {{ editor.currentSceneIndex }}
 			i-input(
-				v-model="editor.sceneIndex",
+				:value="editor.currentSceneIndex",
 				search,
 				readonly,
 				enter-button="复制",
@@ -66,7 +70,9 @@ export default class DLeftScene extends Vue {
 	editor: Editor = Editor.Instance()
 	editScene = false
 	copyModel = false
-
+	changeSceneIndex(index: string | number): void {
+		this.editor.selectSceneIndex(index)
+	}
 	handleSetScene(name: string): void {
 		switch (name) {
 			case 'create':
@@ -76,7 +82,13 @@ export default class DLeftScene extends Vue {
 				this.editScene = true
 				break
 			case 'destroy':
-				this.editor.destroyScene()
+				this.$Modal.confirm({
+					title: '是否删除当前场景？',
+					content: '该场景未删除的组件将自动进入回收站！',
+					onOk: () => {
+						this.editor.destroyScene()
+					},
+				})
 				break
 			case 'copy':
 				this.copyModel = true
@@ -85,7 +97,7 @@ export default class DLeftScene extends Vue {
 	}
 
 	handleCopy(): void {
-		copyText(this.editor.sceneIndex)
+		copyText(this.editor.currentSceneIndex)
 	}
 
 	handleSceneName(e): void {
